[PRED-267] 2스테이지 아트리소스 및 창 크기 변경 커밋

+++ NailEngine/Struct.h
@@ -235,7 +235,7 @@ struct ParticleBuffer
 
 struct LightMapUV
 {
-	int lightMapIndex;
+	float lightMapIndex;
 	DirectX::SimpleMath::Vector3 padding;
 	DirectX::SimpleMath::Vector2 uvOffset;
 	DirectX::SimpleMath::Vector2 scaling;
